Rename modules: mimi_transformer -> transformer, transformer -> attention
