+++ pocket_tts/models/flow_lm.py
@@ -7,7 +7,7 @@ from torch import nn
 from typing_extensions import Self
 
 from pocket_tts.conditioners.text import LUTConditioner
-from pocket_tts.modules.mimi_transformer import StreamingTransformer
+from pocket_tts.modules.transformer import StreamingTransformer
 from pocket_tts.modules.mlp import SimpleMLPAdaLN
 from pocket_tts.utils.config import FlowLMConfig
